% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{AddTrailingSlash}
\alias{AddTrailingSlash}
\title{AddTrailingSlash}
\usage{
AddTrailingSlash(string = "stairway/to/heaven")
}
\arguments{
\item{string}{The file path potentially missing the trailing slash}
}
\description{
Adds a final slash '/', if missing from a string (file path).
}
\examples{
AddTrailingSlash (string = "stairway/to/heaven")
}
