% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{any_print}
\alias{any_print}
\title{any_print}
\usage{
any_print(...)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}
}
\description{
A more flexible printing function that collapses any variable passed to it by white spaces.
}
\examples{
any_print ("Hello ", "you ", 3, ", " , 11, " year old kids.")
}

