% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{as.factor.numeric}
\alias{as.factor.numeric}
\title{as.factor.numeric}
\usage{
as.factor.numeric(vec, rename = FALSE, ...)
}
\arguments{
\item{vec}{vector of factors or strings}

\item{rename}{Rename the vector?}

\item{...}{Pass any other argument to as.factor()}
}
\description{
Turn any vector into numeric categories as.numeric(as.factor(vec))
}
\examples{
as.factor.numeric(LETTERS[1:4])
}
