% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{barplot.label}
\alias{barplot.label}
\title{barplot.label}
\usage{
\method{barplot}{label}(x, y, labels, bottom = F, relpos_top = 0.9,
  relpos_bottom = 0.1, ...)
}
\arguments{
\item{x}{X-position of the label.}

\item{y}{Y-position of the label.}

\item{labels}{Label text.}

\item{bottom}{Put labels at the bottom of the bars.}

\item{relpos_top}{Relative offset.}

\item{relpos_bottom}{Relative offset.}

\item{...}{Pass any other parameter of the corresponding text function (most of them should work).}
}
\description{
Add extra labels to your bar plots at the top or the base.
}
\examples{
barplot.label (x =  , y =  , labels =  , bottom = F, relpos_top = 0.9, relpos_bottom = 0.1, ... =  )
}

