% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{check_OutDir}
\alias{check_OutDir}
\title{Compare Current Working Directory with OutDir}
\usage{
check_OutDir()
}
\arguments{
\item{OutDir}{The target directory to compare with the current working directory.
Default: OutDir (assumes that OutDir should be defined in the global environment).}
}
\value{
Prints a message to the console if the current working directory does not
match OutDir. No return value.
}
\description{
This function checks if the current working directory is the same as a
specified directory (OutDir). If they do not match, it prints both directories to the
screen.
}
\examples{
check_OutDir()

}
