% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{color_check}
\alias{color_check}
\title{color_check}
\usage{
color_check(..., incrBottMarginBy = 0, savefile = FALSE)
}
\arguments{
\item{...}{Additional parameters.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot.}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}
}
\description{
Display the colors encoded by the numbers / color-ID-s you pass on to this function
}
\examples{
color_check(1:3)
}
