% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{create_set_Original_OutDir}
\alias{create_set_Original_OutDir}
\title{Closing pair of the create_set_SubDir function. Call when chaning back to the main dir. Set the output directory of the script, and set the "NewOutDir" variable that is used by all ~wplot functions.}
\usage{
create_set_Original_OutDir(NewOutDir = OutDirOrig, b.Subdirname_ = FALSE,
  setDir = T)
}
\arguments{
\item{NewOutDir}{The new OutDir}

\item{setDir}{Change working directory to the newly defined subdirectory}

\item{NewPath_of_report}{The new path_of_report}
}
\description{
Closing pair of the create_set_SubDir function. Call when chaning back to the main dir. Set the output directory of the script, and set the "NewOutDir" variable that is used by all ~wplot functions.
}
\examples{
create_set_Original_OutDir (...)
}
