% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{cv}
\alias{cv}
\title{cv}
\usage{
cv(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector with numbers}

\item{na.rm}{Remove NA-s? Default: TRUE}
}
\description{
Calculates the coefficient of variation (CV) for a numeric vector (it excludes NA-s by default)
}
\examples{
cv(rnorm(100, sd = 10))
}
