% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{.getScriptName}
\alias{.getScriptName}
\title{Get Current Script Name or Basename of Output Directory}
\usage{
.getScriptName()
}
\value{
A string containing the basename of the current script or the basename of \code{OutDir}
if the script name is unavailable.
}
\description{
This function attempts to retrieve the name of the currently opened script in
the RStudio editor. If the script name cannot be obtained or if the \code{rstudioapi} package is
not available, it returns the basename of the directory specified by \code{OutDir}.
}
