% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{filter_HP}
\alias{filter_HP}
\title{filter_HP}
\usage{
filter_HP(
  numeric_vector,
  threshold,
  passequal = FALSE,
  prepend = "",
  return_survival_ratio = FALSE,
  na.rm = TRUE,
  plot.hist = TRUE,
  saveplot = FALSE,
  ...
)
}
\arguments{
\item{numeric_vector}{Values to be filtered.}

\item{threshold}{A numeric value above which "numeric_vector" passes.}

\item{passequal}{Pass if a value is larger, or equal than the threshold. FALSE by default.}

\item{prepend}{Text prepended to the results.}

\item{return_survival_ratio}{Return a number with the survival ratio (TRUE),
or a logical index vector of the survivors (FALSE).}

\item{na.rm}{Remove NA-s? Default: TRUE}

\item{plot.hist}{Plot the histogram of the input data}

\item{saveplot}{Save the histogram as PDF, FALSE by defeault}

\item{...}{Additional arguments for the histogram}
}
\description{
Filter values that fall between above high-pass-threshold (X >).
}
\examples{
filter_HP (numeric_vector = rnorm(1000, 6), threshold = 5,
 prepend = "From all values ", return_survival_ratio = FALSE)
}
