% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{filter_MidPass}
\alias{filter_MidPass}
\title{filter_MidPass}
\usage{
filter_MidPass(numeric_vector, HP_threshold, LP_threshold, prepend = "",
  return_survival_ratio = FALSE, EdgePass = F, na_rm = T)
}
\arguments{
\item{numeric_vector}{Values to be filtered.}

\item{HP_threshold}{Lower threshold value. (>=)}

\item{LP_threshold}{Upper threshold value. (<)}

\item{prepend}{Text prepended to the results.}

\item{return_survival_ratio}{Return a number with the survival ratio (TRUE), or a logical index vector of the survivors (FALSE).}

\item{EdgePass}{If TRUE, it reverses the filter: everything passes except between the two thresholds.}
}
\description{
Filter values that fall above high-pass-threshold !(X >=)! and below the low-pass threshold (X <).
}
\examples{
filter_MidPass (numeric_vector =  , HP_threshold =  , LP_threshold =  , prepend =  , return_survival_ratio = FALSE, EdgePass = F)
}
