% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{FixUnderscores}
\alias{FixUnderscores}
\title{FixUnderscores}
\usage{
FixUnderscores(string = "stairway__to_heaven_", trimFinal = TRUE)
}
\arguments{
\item{string}{The file path potentially having Double Slash}

\item{trimFinal}{Remove final undescore?}
}
\description{
FixUnderscores removes multiple consecutive underscores (e.g. '\emph{') from a string, and optionally also removes a final '}'.
}
\examples{
FixUnderscores(string = "stairway//to/heaven")
}
