% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{getCategories}
\alias{getCategories}
\title{getCategories
Extract unique entries with a corresponding name.}
\usage{
getCategories(named_categ_vec)
}
\arguments{
\item{named_categ_vec}{A vector of categories with names. "Uniqueness" in the vector and its name should be the same!!!}
}
\description{
getCategories
Extract unique entries with a corresponding name.
}
\examples{
function(named_categ_vec)
}
