% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{iprint}
\alias{iprint}
\title{iprint

A more intelligent printing function that collapses any variable passed to it by white spaces.}
\usage{
iprint(...)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}
}
\description{
iprint

A more intelligent printing function that collapses any variable passed to it by white spaces.
}
\examples{
iprint ("Hello ", "you ", 3, ", ", 11, " year old kids.")
}
