% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{iround}
\alias{iround}
\title{iround}
\usage{
iround(x, digitz = 3)
}
\arguments{
\item{x}{Unrounded number.}

\item{digitz}{Number of digits to keep. 3 by default.}
}
\description{
Rounds a value to the significant amount of digits. Its a wrapper for signif().
}
\examples{
iround (x = 2.3232, digitz = 3)
}
