% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{llwrite_list}
\alias{llwrite_list}
\title{llwrite_list}
\usage{
llwrite_list(yourlist, printName = "self")
}
\arguments{
\item{yourlist}{your list}

\item{printName}{print header level 4: the name of the list or a custom string}
}
\description{
Print a list object from R, one element per line, into your markdown report
}
\examples{
your_list = list(letters[1:4], 5:9); llwrite_list(your_list)
}
