% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{MarkDown_Table_writer_NamedVector}
\alias{MarkDown_Table_writer_NamedVector}
\title{MarkDown_Table_writer_NamedVector}
\usage{
MarkDown_Table_writer_NamedVector(NamedVector, FullPath = path_of_report,
  percentify = F, title_of_table = NA)
}
\arguments{
\item{NamedVector}{A vector for the table body, with names as table header.}

\item{FullPath}{Full path to the file.}

\item{percentify}{Format numbers [0,1] to percentages 0-100.}

\item{title_of_table}{Title above the table (in the markdown report).}
}
\description{
Take an R vector with names, parse a markdown table from it, and write it to the markdown report, set by "path_of_report".
}
\examples{
MarkDown_Table_writer_NamedVector (NamedVector =  , FullPath = path_of_report, percentify = F, title_of_table = NA)
}
