% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{md.import}
\alias{md.import}
\title{md.import}
\usage{
md.import(from.file, to.file = path_of_report)
}
\arguments{
\item{from.file}{File to be appended at the (current) last line of the report}

\item{to.file}{The report file. Defined as "path_of_report" by default, which is set by the "setup_MarkdownReports" function.}
}
\description{
Import and concatenated an external markdown or text file to the report
}
\examples{
md.import("~/Downloads/yourfile.txt")
}
