% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{md.import.table}
\alias{md.import.table}
\title{md.import.table}
\usage{
md.import.table(
  from.file.table,
  title_of_table,
  has.rownames = TRUE,
  has.colnames = TRUE,
  field.sep = "\\t",
  to.file = path_of_report
)
}
\arguments{
\item{from.file.table}{The *.tsv file to be appended  as table at
the (current) last line of the report.}

\item{title_of_table}{Title above the table (as header 4, in the markdown report).}

\item{has.rownames}{If the first column contains (unique!) rownames.}

\item{has.colnames}{If the first line of the file contains the header, or the column names.}

\item{field.sep}{Field separator in table file. Tab's by default.}

\item{to.file}{The report file. Defined as "path_of_report" by default,
which is set by the "setup_MarkdownReports" function.}
}
\description{
Import a table (.csv, or tab seprated values, .tsv file) and write it
in markdown format to the report.
}
\examples{
x = matrix(1:9,3); utils::write.table(x, sep = "\\t", file = "~/x.tsv");
md.import.table("~/x.tsv")
}
