% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{md.List2Table}
\alias{md.List2Table}
\title{md.List2Table}
\usage{
md.List2Table(
  parameterlist,
  title = "List elements",
  colname2 = "Value",
  maxlen = 20
)
}
\arguments{
\item{parameterlist}{List of Paramters.}

\item{title}{Title of the table.}

\item{colname2}{Name of the 2nd column.}

\item{maxlen}{Maximum length of entries in a parameter list element,.}
}
\description{
Broader variant of md.LogSettingsFromList(). Log the values (col2) from a named (col1) list, in a table format
in the report.
}
\examples{
md.LogSettingsFromList(parameterlist = list("min" = 4, "method" = "pearson", "max" = 10))
}
