% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{md.LogSettingsFromList}
\alias{md.LogSettingsFromList}
\title{md.LogSettingsFromList}
\usage{
md.LogSettingsFromList(parameterlist = p, maxlen = 20)
}
\arguments{
\item{parameterlist}{List of Paramters}
}
\description{
Log the parameters & settings used in the script and stored in a list, in a table format in the report.
}
\examples{
md.LogSettingsFromList(parameterlist = list("min"=4, "method"="pearson", "max"=10))
}
