% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{md.write.as.list}
\alias{md.write.as.list}
\title{md.write.as.list}
\usage{
md.write.as.list(vector = 1:3, h = 4, numbered = FALSE, ...)
}
\arguments{
\item{vector}{Vecot to be wirtten as a list}

\item{h}{Level of header above tl list.}

\item{numbered}{TRUE = Numbered list, FALSE = unordered list is written}

\item{...}{Additional parameters}
}
\description{
Writes a vector as a (numbered) list into the report file.
}
\examples{
md.write.as.list()
}
