% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{modus}
\alias{modus}
\title{modus}
\usage{
modus(x)
}
\arguments{
\item{x}{A vector with numbers}
}
\description{
Calculates the modus of a numeric vector (it excludes NA-s by default)
}
\examples{
modus(c(1, 1, 2, 3, 3, 3, 4, 5)); modus(1:4)
}
