% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{na.omit.strip}
\alias{na.omit.strip}
\title{na.omit.strip}
\usage{
\method{na.omit}{strip}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{Values to filter for NA}

\item{silent}{Silence the data structure coversion warning: anything ->vector}

\item{...}{Pass any other argument to na.omit()}
}
\description{
Omit NA values from a vector and return a clean vector without any spam.
}
\examples{
# na.omit.strip(c(1, 2, 3, NA, NaN, 2))
}
