% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{parFlags2}
\alias{parFlags2}
\title{parFlags2}
\usage{
parFlags2(prefix = ".", ..., pasteflg = T, coll.char = ".",
  coll.char.intra = "_")
}
\arguments{
\item{prefix}{Append something before?}

\item{...}{Paramter variables}

\item{pasteflg}{Boolean: paste the parameters-flags together?}

\item{coll.char}{Separating character between each parameters-flag}

\item{coll.char.intra}{Separating character between parameters and its value}
}
\description{
Create a string from the names of the (boolean) parameters (T or F) of true values. Use it for Suffixing plot names with the parameters that were used for that plot.
}
\examples{
pearson = T; filtered =3; normalized = F; MyPlotname = parFlags2(prefix = "MyPlot" , pearson, filtered, normalized ); MyPlotname
}
