% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ParseFilePath}
\alias{ParseFilePath}
\title{ParseFilePath}
\usage{
ParseFilePath(...)
}
\arguments{
\item{...}{The set of strings (character vectors) to be parsed into a file path, and potentially having Double Slashes, potentially missing a trailing slash.}
}
\description{
ParseFilePath pastes elements by slash, then removes Double Slashes '//' from a string and adds a final '/' if missing from a file path.
}
\examples{
ParseFilePath(string = "stairway///to/heaven")
}
