% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{pdfA4plot_off}
\alias{pdfA4plot_off}
\title{pdfA4plot_off}
\usage{
pdfA4plot_off()
}
\description{
The pair of the "pdfA4plot_on()" function; to finish plotting in the A4 pdf.
}
\examples{
pdfA4plot_on.layout(plotname = "MyA4_w_layout")
hist(rnorm(100))
hist(-rnorm(100))
hist(10 + rnorm(100))
pdfA4plot_off()

}
