% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{percentage_formatter}
\alias{percentage_formatter}
\title{percentage_formatter}
\usage{
percentage_formatter(x, digitz = 3)
}
\arguments{
\item{x}{A vector of numbers between [0, 1]/}

\item{digitz}{Number of digits to keep. 3 by default.}
}
\description{
Parse a string of 0-100% from a number between 0 and 1.
}
\examples{
percentage_formatter (x =  , digitz = 3)
}
