% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{qlegend}
\alias{qlegend}
\title{qlegend
# Quickly add a legend, and save the plot immediately}
\usage{
qlegend(NamedColorVec, poz = 3, ..., w_ = 7, h_ = w_, bty = "n",
  OverwritePrevPDF = T)
}
\arguments{
\item{poz}{1:4 corresponding to "topleft","topright", "bottomright", "bottomleft"}

\item{...}{Additional parameters for legend()}

\item{w_}{Width of the saved pdf image, in inches.}

\item{h_}{Height of the saved pdf image, in inches.}

\item{bty}{The type of box to be drawn around the legend. The allowed values are "o" (the default) and "n".}

\item{OverwritePrevPDF}{Save the plot immediately with the same name the last wplot* function made (It is stored in plotnameLastPlot variable).}
}
\description{
qlegend
# Quickly add a legend, and save the plot immediately
}
\examples{
ccc = as.factor.numeric(c(6,7,8,7,7,6,6,8))  ; qlegend(ccc) # Uses as.factor.numeric() from Github / Vertesy / TheCorvinas
}
