% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{RemoveDoubleSlash}
\alias{RemoveDoubleSlash}
\title{RemoveDoubleSlash}
\usage{
RemoveDoubleSlash(string = "stairway//to/heaven")
}
\arguments{
\item{string}{The file path potentially having Double Slash}
}
\description{
RemoveDoubleSlash removes multiple consecutive slashes (e.g. '//') from a string (file path). Also works for 2,3 consecutive slashes
}
\examples{
RemoveDoubleSlash (string = "stairway//to///heaven")
}
