% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{substrRight}
\alias{substrRight}
\title{substrRight}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{a character vector.}

\item{n}{integer. The number of elements on the right to be kept.}
}
\description{
Take the right substring of a string
}
\examples{
substrRight  ("Not cool", n = 4)
}
