% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{superscript_in_plots}
\alias{superscript_in_plots}
\title{superscript_in_plots}
\usage{
superscript_in_plots(prefix = "n", sup = "k", suffix = "")
}
\arguments{
\item{prefix}{String before the superscript.}

\item{sup}{Superscripted text.}

\item{suffix}{String after the subscript.}
}
\description{
superscript_in_plots
}
\examples{
plot (1, 1, main =superscript_in_plots())
}

