% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{trueUnlessSpec}
\alias{trueUnlessSpec}
\title{trueUnlessSpec}
\usage{
trueUnlessSpec(nameOfaBoolean)
}
\arguments{
\item{nameOfaBoolean}{Name of a possible defined variable to be tested.}
}
\description{
Return TRUE unless the variable is defined. if defined, it returns the variable
}
\examples{
trueUnlessSpec("xsadasf32"); B=F; trueUnlessSpec("F"); trueUnlessSpec("c")
}
