% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{try.dev.off}
\alias{try.dev.off}
\title{try.dev.off

Tries to close R graphical devices without raising an error.}
\usage{
try.dev.off()
}
\description{
try.dev.off

Tries to close R graphical devices without raising an error.
}
\examples{
try.dev.off ()
}
