% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ttl_field}
\alias{ttl_field}
\title{ttl_field}
\usage{
ttl_field(flname = fname)
}
\arguments{
\item{flname}{Name of the plot}
}
\description{
Internal function. Creates the string written into the PDF files "Title' (metadata) field/
}
\examples{
ttl_field()
}
