% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{unless.specified}
\alias{unless.specified}
\title{unless.specified}
\usage{
unless.specified(NameOfaVariable, def = TRUE)
}
\arguments{
\item{NameOfaVariable}{Name of a possibly defined variable to be tested.}

\item{def}{Default return value}
}
\description{
Return value X (TRUE by default) unless the variable is defined.
If defined, it returns the variable.
}
\examples{
unless.specified("xsadasf32", 2); Num = 22; unless.specified("Num", 1); unless.specified("c", 333)
}
