% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{UnlessSpec}
\alias{UnlessSpec}
\title{UnlessSpec}
\usage{
UnlessSpec(NameOfaVariable, def = TRUE)
}
\arguments{
\item{NameOfaVariable}{Name of a possibly defined variable to be tested.}

\item{def}{Default return value}
}
\description{
Return value X (TRUE by default) unless the variable is defined. If defined, it returns the variable.
}
\examples{
UnlessSpec(2,"xsadasf32"); Num=22; UnlessSpec(2,"Num"); UnlessSpec(2, "c")
}
