% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{variable.or.path.exists}
\alias{variable.or.path.exists}
\title{variable.or.path.exists}
\usage{
variable.or.path.exists(path = path_of_report)
}
\arguments{
\item{path}{A variable name that might not exist and might point to a non-existent direcotry.}
}
\description{
Check if a variable name is defined, and if so, does the path (to a file) stored in that variable points to an existing directory?
}
\examples{
variable.or.path.exists (path = path_of_report)
}
