% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{vvenn}
\alias{vvenn}
\title{vvenn}
\usage{
vvenn(yalist, imagetype = "png", alpha = 0.5, ..., w = 7, h = 7,
  mdlink = F)
}
\arguments{
\item{yalist}{The variable to plot.}

\item{imagetype}{Image format, png by default.}

\item{alpha}{Transparency, .5 by default.}

\item{...}{Pass any other parameter of the corresponding venn.diagram() function (most of them should work).}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}
}
\description{
Save venn diagrams. Unlike other ~vplot funcitons, this saves directly into a .png, abnd does not use dev.copy2pdf()
}
\examples{
vvenn (yalist =  , imagetype = png, alpha = 0.5, ... =  , w = 7, h = 7, mdlink = F)
}

