% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{whist}
\alias{whist}
\title{whist}
\usage{
whist(
  variable,
  ...,
  breaks = 20,
  col = get0("b.def.color", ifnotfound = "gold1"),
  plotname = substitute(variable),
  main = kollapse("Histogram of ", substitute(variable)),
  xlab = substitute(variable),
  lty = 2,
  lwd = 3,
  lcol = 1,
  filtercol = 0,
  vline = FALSE,
  filter = c(FALSE, "HighPass", "LowPass", "MidPass")[1],
  passequal = TRUE,
  savefile = get0("b.save.wplots", ifnotfound = TRUE),
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  mdlink = ww.set.mdlink(),
  PNG = get0("b.usepng", ifnotfound = TRUE)
)
}
\arguments{
\item{variable}{The variable to plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}

\item{breaks}{Number of bins.}

\item{col}{Color of the plot.}

\item{plotname}{The name of the file.}

\item{main}{Title of the plot.}

\item{xlab}{X-axis label.}

\item{lty}{Linetype, defined by numbers 1-6.}

\item{lwd}{Line width. Set to 2 by default.}

\item{lcol}{Color of the line.}

\item{filtercol}{Color bars below / above the threshold with red / green. Define the direction by
-1 or 1. Takes effect if "vline" is defined.}

\item{vline}{Draw a vertical line at the value you pass on to it. Useful to display a threshold.
Design the line by "lty", "lwd" & "lcol" parameters.}

\item{filter}{filtervalues}

\item{passequal}{Pass equal values}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}
}
\description{
Create and save histograms as .pdf, in "OutDir". If mdlink = TRUE, it inserts a .pdf and a .png
link in the markdown report, set by "path_of_report". The .png version is not created, only the
link is put in place, not to overwrite previous versions. Name the file by naming the variable!
Cannot be used with dynamically called variables \link{e.g. call vectors within a loop}. "filtercol"
assumes  >= coloring!
}
\examples{
MyGauss <- rnorm(1000)
whist(
  variable = MyGauss, col = "gold1", w = 7,
  breaks = 20, mdlink = FALSE, hline = FALSE, vline = FALSE, lty = 2, lwd = 3,
  lcol = 2, filtercol = 0
)

}
