% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wlegend2}
\alias{wlegend2}
\title{wlegend2}
\usage{
wlegend2(x = "bottomleft", fill_ = NULL, legend = names(fill_), ...,
  w_ = 7, h_ = w_, bty = "n", OverwritePrevPDF = T)
}
\arguments{
\item{x}{}

\item{fill_}{Color of the boxes next to the text}

\item{legend}{Labels displayed (Text)}

\item{...}{Additional parameters for legend()}

\item{w_}{Width of the saved pdf image, in inches.}

\item{h_}{Height of the saved pdf image, in inches.}

\item{bty}{The type of box to be drawn around the legend. The allowed values are "o" (the default) and "n".}

\item{OverwritePrevPDF}{Save the plot immediately with the same name the last wplot* function made (It is stored in plotnameLastPlot variable).}
}
\description{
wlegend2
}
\examples{
function(x="bottomleft", fill_ = NULL, legend = names(fill_), ..., w_=7, h_=w_, bty = "n", OverwritePrevPDF =T)
}
