% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wplot_save_pheatmap}
\alias{wplot_save_pheatmap}
\title{wplot_save_pheatmap}
\usage{
wplot_save_pheatmap(
  x,
  suffix = NULL,
  filename = kpp(substitute(x), suffix),
  width = 15,
  height = width
)
}
\arguments{
\item{x}{The pheatmap object to save.}

\item{suffix}{Suffix to File name.}

\item{filename}{File name (saved as .pdf, inside working directory).}

\item{width}{width of the plot in inches.}

\item{height}{height of the plot in inches.}
}
\description{
Save pheatmap object. Modified from:
https://stackoverflow.com/questions/43051525/how-to-draw-pheatmap-plot-to-screen-and-also-save-to-file
}
\examples{
test = matrix(rnorm(200), 20, 10);
colnames(test) = paste("Test", 1:10, sep = "");
rownames(test) = paste("Gene", 1:20, sep = "");
ph.test <- pheatmap::pheatmap(test);
wplot_save_pheatmap(ph.test)
}
