% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wplot_save_this}
\alias{wplot_save_this}
\title{wplot_save_this}
\usage{
wplot_save_this(
  plotname = MarkdownHelpers::ww.autoPlotName(),
  OverwritePrevPDF = TRUE,
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  mdlink = FALSE,
  PNG = get0("b.usepng", ifnotfound = FALSE),
  ...
)
}
\arguments{
\item{plotname}{Title of the plot (main parameter) and also the name of the file.}

\item{OverwritePrevPDF}{Overwrite previous PDF image (as name stored in plotnameLastPlot).
If FALSE, it creates a name from the date.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}
}
\description{
Save the currently active graphic device (for complicated plots).  Insert links to your markdown
report, set by "path_of_report". Name the file by naming the variable!
}
\examples{
wplot_save_this(
  plotname = date(), col = "gold1", w = 7,
  mdlink = FALSE, ManualName = FALSE
)
}
