% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{write.simple.tsv}
\alias{write.simple.tsv}
\title{write.simple.tsv}
\usage{
write.simple.tsv(
  input_df,
  extension = "tsv",
  ManualName = "",
  o = FALSE,
  gzip = FALSE,
  ...
)
}
\arguments{
\item{input_df}{Your Dataframe with row- and column-names}

\item{extension}{e.g.: tsv}

\item{ManualName}{Specify full filename if you do not want to name it by the variable name.}

\item{o}{Open the file after saving? FALSE by default}

\item{gzip}{Compress the file after saving? FALSE by default}

\item{...}{Pass any other argument to the kollapse() function used for file name.}
}
\description{
Write out a matrix-like R-object WITH ROW- AND COLUMN- NAMES to a file with as tab separated
values (.tsv). Your output filename will be either the variable's name. The output file will be
located in "OutDir" specified by you at the beginning of the script, or under your current
working directory. You can pass the PATH and VARIABLE separately (in order), they will be
concatenated to the filename.
}
\examples{
YourDataFrameWithRowAndColumnNames = cbind("A" = rnorm(100), "B" = rpois(100, 8))
rownames(YourDataFrameWithRowAndColumnNames) = letters[1:NROW(YourDataFrameWithRowAndColumnNames)]
write.simple.tsv(YourDataFrameWithRowAndColumnNames)
}
