% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wstripchart_list}
\alias{wstripchart_list}
\title{wstripchart_list}
\usage{
wstripchart_list(
  yourlist,
  ...,
  main = as.character(substitute(yourlist)),
  sub = NULL,
  ylab = "",
  xlab = "",
  border = 1,
  bxpcol = 0,
  pch = 18,
  pchlwd = 1,
  pchcex = 1.5,
  incrBottMarginBy = 0,
  tilted_text = FALSE,
  bg = "chartreuse2",
  col = "black",
  method = "jitter",
  jitter = 0.2,
  savefile = get0("b.save.wplots", ifnotfound = FALSE),
  w = get0("b.defSize", ifnotfound = TRUE),
  h = w,
  mdlink = MarkdownHelpers::ww.set.mdlink(),
  PNG = get0("b.usepng", ifnotfound = FALSE)
)
}
\arguments{
\item{yourlist}{Input list to be plotted.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}

\item{main}{Title of the plot (main parameter) and also the name of the file.}

\item{sub}{Subtitle below the plot.}

\item{ylab}{Y-axis label.}

\item{xlab}{X-axis label.}

\item{border}{An optional vector of colors for the outlines of the boxplots. The values in border
are recycled if the length of border is less than the number of plots.}

\item{bxpcol}{Color of the boxplot outlines.}

\item{pch}{Define the symbol for each data point. A number (0-25) or any string between ""-s.}

\item{pchlwd}{Define the outline width of the symbol for each data point.}

\item{pchcex}{Define the size of the symbol for each data point.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not
fit on the plot.}

\item{tilted_text}{Use 45 degree x-labels if TRUE. Useful for long, but not too many labels.}

\item{bg}{Background color.}

\item{col}{Color of the plot.}

\item{method}{Method for displaying data points to avoid overlap; either "jitter" or "stack".
See stripchart().}

\item{jitter}{The amount of horizontal scatter added to the individual data points (to avoid
overlaps).}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}
}
\description{
Create and save stripcharts from a list as .pdf, in "OutDir". This version allows individual
coloring of each data point, by a color-list of the same dimension. If mdlink = TRUE, it inserts a
.pdf and a .png link in the markdown report, set by "path_of_report". The .png version is not
created, only the link is put in place, not to overwrite previous versions.
}
\examples{
try.dev.off()
my.ls <- list(A = rnorm(10), B = rnorm(10), C = rnorm(10))
wstripchart_list(
  yourlist = my.ls, sub = NULL, ylab = NULL, xlab = NULL,
  border = 1, bxpcol = 0, pch = 23, pchlwd = 1, pchcex = 1.5, bg = "chartreuse2", col = 1,
  method = jitter, jitter = 0.2, w = 7, incrBottMarginBy = 0, tilted_text = FALSE, mdlink = FALSE
)
}
