% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wvenn}
\alias{wvenn}
\title{wvenn}
\usage{
wvenn(
  yourlist,
  imagetype = "png",
  alpha = 0.5,
  fill = seq_along(yourlist),
  subt,
  ...,
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  mdlink = MarkdownHelpers::ww.set.mdlink(),
  plotname = substitute(yourlist),
  openFolder = TRUE
)
}
\arguments{
\item{yourlist}{The variable to plot.}

\item{imagetype}{Image format, png by default.}

\item{alpha}{Transparency, .5 by default.}

\item{fill}{Background color vec}

\item{subt}{Subtitle}

\item{...}{Pass any other parameter of the corresponding venn.diagram()
function (most of them should work).}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report,
set by "path_of_report".}

\item{plotname}{Manual plotname parameter}

\item{openFolder}{open current directory (=working if setup_MarkdownReports('setDir=TRUE'))}
}
\description{
Save Venn diagrams. Unlike other ~vplot functions, this saves directly into a .png,
and it does not use the dev.copy2pdf() function.
}
\examples{
TwoSets <- list("set1" = LETTERS[1:6], "set2" = LETTERS[3:9])
wvenn(yourlist = TwoSets, imagetype = "png", alpha = 0.5, w = 7, mdlink = FALSE)
}
