% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wvioplot_list}
\alias{wvioplot_list}
\title{wvioplot_list}
\usage{
wvioplot_list(
  yourlist,
  ...,
  main = as.character(substitute(yourlist)),
  sub = NULL,
  xlab = names(yourlist),
  ylab = "",
  ylim = FALSE,
  col = c(2:(length(yourlist) + 1)),
  incrBottMarginBy = 0,
  tilted_text = FALSE,
  yoffset = 0,
  savefile = get0("b.save.wplots", ifnotfound = FALSE),
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  mdlink = MarkdownHelpers::ww.set.mdlink(),
  PNG = get0("b.usepng", ifnotfound = FALSE)
)
}
\arguments{
\item{yourlist}{Input list to be plotted.}

\item{...}{Pass any other parameter of the corresponding
plotting function (most of them should work).}

\item{main}{Title of the plot (main parameter) and also the name of the file.}

\item{sub}{Subtitle below the plot.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{ylim}{Manual y axis limits}

\item{col}{Color of the plot.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot.
Use if labels do not fit on the plot.}

\item{tilted_text}{Use 45 degree x-labels if TRUE. Useful for long, but not too many labels.}

\item{yoffset}{Offset for X axis labels (in vertical / Y dimension).}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report,
set by "path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}
}
\description{
Create and save violin plots as .pdf, in "OutDir". It requires (and calls) "vioplot" package.
If mdlink = TRUE, it inserts a .pdf and a .png link in the markdown report,
set by "path_of_report". The .png version is not created, only the link is put in place,
not to overwrite previous versions.
}
\examples{
try.dev.off()
my.ls <- list(A = rnorm(10), B = rnorm(10), C = rnorm(10))
# wvioplot_list (yourlist = my.ls, xlab = names(yourlist), ylab = "", incrBottMarginBy = 0,
# w = 7, tilted_text = FALSE, mdlink = FALSE)
}
