% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.assign_to_global}
\alias{ww.assign_to_global}
\title{ww.assign_to_global}
\usage{
ww.assign_to_global(name, value, pos = 1)
}
\arguments{
\item{name}{Name of the global variabe to be assigned}

\item{value}{Value of the global variabe to be assigned}

\item{pos}{defaults to 1 which equals an assingment to global environment}
}
\description{
function loading results in global environment.
Source: https://stackoverflow.com/questions/28180989/
}
