% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.dev.copy}
\alias{ww.dev.copy}
\title{ww.dev.copy}
\usage{
ww.dev.copy(PNG_ = F, w_ = w, h_ = h, fname_ = fname)
}
\arguments{
\item{PNG_}{Set to true if you want to save the plot as PNG instead of the default PDF.}

\item{w_}{Width of the saved pdf image, in inches.}

\item{h_}{Height of the saved pdf image, in inches.}

\item{fname_}{File name}
}
\description{
Parser for dev.copy to save as PDF or PNG
}
\examples{
ww.dev.copy(PNG=F, w_=7, h_=7, fname_ = "fname")
}
