% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.set.path_of_report}
\alias{ww.set.path_of_report}
\title{ww.set.path_of_report}
\usage{
ww.set.path_of_report()
}
\description{
Checks if global variable path_of_report is defined. If not,
it defines it as Analysis.md in the current working directory
}
\examples{
ww.set.path_of_report()
}
