% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.ttl_field}
\alias{ww.ttl_field}
\title{ww.ttl_field}
\usage{
ww.ttl_field(plotname, creator = "MarkdownReports")
}
\arguments{
\item{plotname}{Name of the plot}

\item{creator}{String X in: "plotblabla by X". Defaults: "MarkdownReports".}
}
\description{
Internal function. Creates the string written into the PDF files "Title' (metadata) field.
}
\examples{
ww.ttl_field("/Users/myplot.jpg")
}
