% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.variable.exists.and.true}
\alias{ww.variable.exists.and.true}
\title{ww.variable.exists.and.true}
\usage{
ww.variable.exists.and.true(var, alt.message = NULL)
}
\arguments{
\item{var}{A variable}

\item{alt.message}{Alternative message if the variable + path does not exist. FALSE or string.}
}
\description{
Check if a variable name is defined, and if so, is it TRUE
}
\examples{
ww.variable.and.path.exists(path = B, alt.message = "Hello, your path/var does not exist.")
}
