% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{barplot_label}
\alias{barplot_label}
\title{barplot_label}
\usage{
barplot_label(
  barplotted_variable,
  labels = iround(barplotted_variable),
  bottom = FALSE,
  TopOffset = 0.5,
  relpos_bottom = 0.1,
  OverwritePrevPDF = unless.specified("b.save.wplots"),
  filename = plotnameLastPlot,
  PNG_ = unless.specified("b.usepng", FALSE),
  w = 7,
  h = w,
  ...
)
}
\arguments{
\item{barplotted_variable}{The variable that you previously bar plotted.}

\item{labels}{Label text.}

\item{bottom}{Put labels at the bottom of the bars.}

\item{TopOffset}{Absolute offset from the top.}

\item{relpos_bottom}{Relative offset from the bottom.}

\item{OverwritePrevPDF}{Save the plot immediately with the same name the last
wplot* function made (It is stored in the plotnameLastPlot variable). Never inserts an mdlink.}

\item{filename}{Filename to overwrite after error bars are added to the current bar plot.}

\item{PNG_}{Set to true if you want to save the plot as PNG instead of the default PDF.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{...}{Pass any other parameter of the corresponding
text function (most of them should work).}
}
\description{
Add extra labels to your bar plots at the top or the base.
}
\examples{
barplot(1:10)
barplot_label(barplotted_variable = 1:10, labels = 11:2, filename = "myBarplot.pdf")
}
