% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{barplot_label}
\alias{barplot_label}
\title{barplot_label}
\usage{
barplot_label(barplotted_variable, labels, bottom = F, TopOffset = 0.5,
  relpos_bottom = 0.1, OverwritePrevPDF = UnlessSpec("b.save.wplots"), ...)
}
\arguments{
\item{barplotted_variable}{The variable that you barplotted previously.}

\item{labels}{Label text.}

\item{bottom}{Put labels at the bottom of the bars.}

\item{TopOffset}{Absolute offset from top.}

\item{relpos_bottom}{Relative offset from bottom.}

\item{OverwritePrevPDF}{Save the plot immediately with the same name the last wplot* function made (It is stored in plotnameLastPlot variable). Never inserts an mdlink.}

\item{...}{Pass any other parameter of the corresponding text function (most of them should work).}
}
\description{
Add extra labels to your bar plots at the top or the base.
}
\examples{
barplot (1:10); barplot_label (1:10, labels =11:2  , bottom = F, TopOffset = 0.5, relpos_bottom = 0.1, ... =  )
}
