% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{continue_logging_markdown}
\alias{continue_logging_markdown}
\title{continue_logging_markdown}
\usage{
continue_logging_markdown(b.scriptname)
}
\arguments{
\item{b.scriptname}{Name of the report file.}
}
\description{
Continue writing to an existing report file.
}
\examples{
OutDir <- paste0(getwd(), "/", collapse = "")
continue_logging_markdown(b.scriptname = "Analysis")
}
