% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{create_set_OutDir}
\alias{create_set_OutDir}
\title{create_set_OutDir}
\usage{
create_set_OutDir(
  ...,
  setDir = TRUE,
  writeScriptPath = TRUE,
  ScrPath = "__corresponding.R.script",
  verbose = TRUE,
  newName = NULL
)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed consecutively.}

\item{setDir}{Set the working directory to OutDir? Default: TRUE}

\item{writeScriptPath}{Make it easier to trace the origin of the output.
Write the path to the current R script into a file in the output directory? Default: TRUE}

\item{ScrPath}{File name in the OutDir. Default: "__corresponding.R.script"}

\item{verbose}{Print directory to screen? Default: TRUE}

\item{newName}{Create a new variable with same path as the "OutDir" variable. Useful since
OutDir may be redefined by other scripts}
}
\description{
Create and set (setwd) the output directory of the script, and define the "OutDir" variable
as a global variable. OutDir is used by most @vertesy and all ~wplot functions. It also writes the path of
the current R script
}
\examples{
create_set_OutDir(setDir = TRUE, getwd())
}
