% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{create_set_OutDir}
\alias{create_set_OutDir}
\title{create_set_OutDir}
\usage{
create_set_OutDir(..., setDir = TRUE, verbose = TRUE)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}

\item{setDir}{Set the working directory to OutDir? Default: TRUE}

\item{verbose}{Print directory to screen? Default: TRUE}
}
\description{
Create or set the output directory of the script, and set the "OutDir" variable that is used by
all ~wplot functions.
}
\examples{
create_set_OutDir(setDir = TRUE, getwd(), "/")
}
