% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{create_set_SubDir}
\alias{create_set_SubDir}
\title{Create or set the output directory of the script, and set the "NewOutDir" variable that is used by all ~wplot functions.}
\usage{
create_set_SubDir(..., makeOutDirOrig = T, setDir = T)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}

\item{makeOutDirOrig}{Change the "OutDirOrig" variable to the current OutDir (before setting it to a subdir).}

\item{setDir}{Change working directory to the newly defined subdirectory}
}
\description{
Create or set the output directory of the script, and set the "NewOutDir" variable that is used by all ~wplot functions.
}
\examples{
create_set_NewOutDir (...)
}
