% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{filter_HP}
\alias{filter_HP}
\title{filter_HP}
\usage{
filter_HP(numeric_vector, threshold, prepend = "",
  return_survival_ratio = F)
}
\arguments{
\item{numeric_vector}{Values to be filtered.}

\item{threshold}{A numeric value above which "numeric_vector" passes.}

\item{prepend}{Text prepended to the results.}

\item{return_survival_ratio}{Return a number with the survival ratio (TRUE), or a logical index vector of the survivors (FALSE).}
}
\description{
Filter values that fall between above high-pass-threshold (X >).
}
\examples{
filter_HP (numeric_vector =  , threshold =  , prepend =  , return_survival_ratio = F)
}

