% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{filter_LP}
\alias{filter_LP}
\title{filter_LP}
\usage{
filter_LP(numeric_vector, threshold, passequal = F, prepend = "",
  return_survival_ratio = F)
}
\arguments{
\item{numeric_vector}{Values to be filtered.}

\item{threshold}{A numeric value below which "numeric_vector" passes.}

\item{passequal}{Pass if a value is smaller, or equal than the threshold. FALSE by default.}

\item{prepend}{Text prepended to the results.}

\item{return_survival_ratio}{Return a number with the survival ratio (TRUE), or a logical index vector of the survivors (FALSE).}
}
\description{
Filter values that fall below the low-pass threshold (X <).
}
\examples{
filter_LP (numeric_vector =  , threshold =  , prepend =  , return_survival_ratio = F)
}

