% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{kollapse}
\alias{kollapse}
\title{kollapse}
\usage{
kollapse(..., print = T)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}

\item{print}{Print the results to the terminal. TRUE by default.}
}
\description{
Collapses values and strings to one string (without a white space). It also prints the results (good for a quick check)
}
\examples{
kollapse("Hello ", "you ", 3, ", " , 11, " year old kids.")
}

