% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{llwrite_list}
\alias{llwrite_list}
\title{llwrite_list
Print a list, one element per line,  into your markdown report}
\usage{
llwrite_list(yalist)
}
\arguments{
\item{yalist}{your list}
}
\description{
llwrite_list
Print a list, one element per line,  into your markdown report
}
\examples{
llwrite_list(your_list)
}
