% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{log_settings_MarkDown}
\alias{log_settings_MarkDown}
\title{log_settings_MarkDown}
\usage{
log_settings_MarkDown(...)
}
\arguments{
\item{...}{Variables (strings, vectors) to be collapsed in consecutively.}
}
\description{
Log the parameters & settings used in the script in a table format.
}
\examples{
log_settings_MarkDown (... =  )
}

