% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{modus}
\alias{modus}
\title{modus}
\usage{
modus(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Calculates the mode (modus) of a numeric vector (it excludes NA-s by default). https://en.wikipedia.org/wiki/Mode_(statistics)
}
\examples{
modus(c(1, 1, 2, 3, 3, 3, 4, 5)); modus(1:4)
}
