% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{pdfA4plot_on.layout}
\alias{pdfA4plot_on.layout}
\title{pdfA4plot_on.layout
Create A4 PDFs to plot multiple subplots in one file with custom numbers of columns in each row}
\usage{
pdfA4plot_on.layout(pname = date(), ..., w = 8.27, h = 11.69,
  layout_mat = rbind(1, c(2, 3), 4:5), mdlink = FALSE,
  title = ttl_field(pname))
}
\arguments{
\item{pname}{Title of the plot (main parameter) and also the name of the file.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{layout_mat}{A matrix of plot layout. Default: rbind(1, c(2, 3), 4:5)}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{title}{Manually set the title field of the PDF file}
}
\description{
pdfA4plot_on.layout
Create A4 PDFs to plot multiple subplots in one file with custom numbers of columns in each row
}
\examples{
pdfA4plot_on.layout();  hist(rnorm(100)); hist(-rnorm(100)); hist(10+rnorm(100)); pdfA4plot_off()
}
