% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{pdfA4plot_on}
\alias{pdfA4plot_on}
\title{pdfA4plot_on}
\usage{
pdfA4plot_on(pname = date(), ..., w = UnlessSpec(8.27,
  "b.defSize.fullpage"), h = 11.69, rows = 4, cols = rows - 1,
  mdlink = ww.set.mdlink(), title = ww.ttl_field(pname))
}
\arguments{
\item{pname}{Title of the plot (main parameter) and also the name of the file.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{w}{Width of the saved pdf image, in inches. c("A4" = 8.27, "1col.nature" = 3.50, "2col.nature" = 7.20, "1col.cell" = 3.35, "1.5col.cell" = 4.49, "2col.cell" = 6.85).}

\item{h}{Height of the saved pdf image, in inches.}

\item{rows}{Number of rows for subplots}

\item{cols}{Number of columns for subplots}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{title}{Manually set the title field of the PDF file}
}
\description{
Create A4 PDFs to plot multiple subplots in one file
}
\examples{
pdfA4plot_on();  hist(rnorm(100)); hist(-rnorm(100)); hist(10+rnorm(100)); pdfA4plot_off()
}
