% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{pdfA4plot_on}
\alias{pdfA4plot_on}
\title{pdfA4plot_on}
\usage{
pdfA4plot_on(pname = date(), ..., w = 8.27, h = 11.69, rows = 4,
  cols = rows - 1, mdlink = FALSE, title = ttl_field(pname))
}
\arguments{
\item{pname}{Title of the plot (main parameter) and also the name of the file.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{rows}{Number of rows for subplots}

\item{cols}{Number of columns for subplots}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{title}{Manually set the title field of the PDF file}
}
\description{
Create A4 PDFs to plot multiple subplots in one file
}
\examples{
pdfA4plot_on();  hist(rnorm(100)); hist(-rnorm(100)); hist(10+rnorm(100)); pdfA4plot_off()
}
