% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{setup_MarkdownReports}
\alias{setup_MarkdownReports}
\title{Setup Markdown Reports}
\usage{
setup_MarkdownReports(
  OutDir = getwd(),
  scriptname = .getScriptName(),
  title = "",
  setDir = TRUE,
  newName = NULL,
  recursive.folder = TRUE,
  backupfolder = TRUE,
  append = FALSE,
  addTableOfContents = FALSE,
  saveSessionInfo = TRUE,
  saveParameterList = "p",
  b.defSize = c(def = 7, A4 = 8.27, `1col.nature` = 3.5, `2col.nature` = 7.2, `1col.cell`
    = 3.35, `1.5col.cell` = 4.49, `2col.cell` = 6.85)[1],
  b.defSize.fullpage = 8.27,
  b.usepng = FALSE,
  b.png4Github = FALSE,
  b.mdlink = TRUE,
  b.save.wplots = TRUE,
  b.def.color = "gold1"
)
}
\arguments{
\item{OutDir}{The output directory (absolute/full path).}

\item{scriptname}{Name of the script file being run, used in the title field
of .pdf files.}

\item{title}{Optional manual title for the report.}

\item{setDir}{Logical, whether to set the working directory to OutDir.
Default is TRUE.}

\item{newName}{Optional new variable with the same path as the "OutDir"
variable, useful if "OutDir" is redefined by other scripts.}

\item{recursive.folder}{Logical, whether to create the output folder
recursively. Default is FALSE.}

\item{backupfolder}{Logical, whether to create a timestamped backup folder
inside OutDir.}

\item{append}{Logical, set to TRUE to avoid overwriting the previous report.}

\item{addTableOfContents}{Logical, whether to add a Table Of Contents in the
report.}

\item{saveSessionInfo}{Logical, whether to save session info to a file.}

\item{saveParameterList}{Logical or character, whether to save the list of
parameters from the given variable (default "p") as a table in the report.}

\item{b.defSize}{Default width of plot, except in pdfA4plot_on(). Default is
7 inches.}

\item{b.defSize.fullpage}{Default width of plot in pdfA4plot_on(). Default is
8.27 inches.}

\item{b.usepng}{Logical, whether to create links to .png versions of plots.}

\item{b.png4Github}{Logical, whether to create GitHub compatible links to
.png plots.}

\item{b.mdlink}{Logical, whether to link saved (.pdf) plots in the report.}

\item{b.save.wplots}{Logical, whether to save plots as .pdf files.}

\item{b.def.color}{Default color for all wplot functions. Default is "gold1".}
}
\description{
Initializes the markdown report file and the output directory.
This function creates a subdirectory in "OutDir", named with the script name
and the modification time. It also defines the "path_of_report" variable
used by all log-writing and wplot functions.
}
\examples{
setup_MarkdownReports(
  OutDir = "path/to/output",
  scriptname = "MyRscript.R",
  title = "Awesome Analysis",
  append = TRUE,
  b.png4Github = TRUE
)
}
