% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{subscript_in_plots}
\alias{subscript_in_plots}
\title{subscript_in_plots}
\usage{
subscript_in_plots(prefix = "log", subscr = 2,
  quantity = "arbitrary units")
}
\arguments{
\item{prefix}{String before the subscript.}

\item{subscr}{Subscripted text.}

\item{quantity}{String in brackets after the subscript, eg.: log2(read count).}
}
\description{
subscript_in_plots
}
\examples{
plot (1, 1, xlab =subscript_in_plots(subscr = 10,quantity = "read count"), ylab =subscript_in_plots())
}

