% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{superscript_in_plots}
\alias{superscript_in_plots}
\title{superscript_in_plots}
\usage{
superscript_in_plots(prefix = "n", sup = "k", suffix = "")
}
\arguments{
\item{prefix}{String before the superscript.}

\item{sup}{Superscripted text.}

\item{suffix}{String after the subscript.}
}
\description{
Returns a formatted string that you feed to main, xlab or ylab parameters of a plot
Create an expression with superscript for axis labels.
Parsed when provided to xlab or ylab of a function.
}
\examples{
plot(1, 1, main = superscript_in_plots())
}
