% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wbarplot}
\alias{wbarplot}
\title{wbarplot}
\usage{
wbarplot(
  variable,
  ...,
  col = get0("b.def.color", ifnotfound = "gold1"),
  sub = FALSE,
  plotname = substitute(variable),
  main = plotname,
  tilted_text = FALSE,
  ylim = NULL,
  hline = FALSE,
  vline = FALSE,
  filtercol = 1,
  lty = 1,
  lwd = 2,
  lcol = 2,
  errorbar = FALSE,
  upper = 0,
  lower = upper,
  arrow_width = 0.1,
  arrow_lwd = 1,
  savefile = get0("b.save.wplots", ifnotfound = TRUE),
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  incrBottMarginBy = 0,
  mdlink = MarkdownHelpers::ww.set.mdlink(),
  PNG = get0("b.usepng", ifnotfound = FALSE)
)
}
\arguments{
\item{variable}{The variable to plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}

\item{col}{Color of the plot.}

\item{sub}{Subtitle below the plot.}

\item{plotname}{The name of the file saved.}

\item{main}{The title of the plot.}

\item{tilted_text}{Use 45 degree x-labels if TRUE. Useful for long, but not too many labels.}

\item{ylim}{Defines the Y axis range. Replacement for the standard "ylim" argument.}

\item{hline}{Draw a horizontal line at the value you pass on to it. Useful to display a
threshold. Design the line by "lty", "lwd" & "lcol" parameters.}

\item{vline}{Draw a vertical line at the value you pass on to it. Useful to display a threshold.
Design the line by "lty", "lwd" & "lcol" parameters.}

\item{filtercol}{Color bars below / above the threshold with red / green. Define the direction by
-1 or 1. Takes effect if "hline" is defined.}

\item{lty}{Linetype, defined by numbers 1-6.}

\item{lwd}{Linewidth. Set to 2 by default.}

\item{lcol}{Color of the line.}

\item{errorbar}{Draw error bars if TRUE. Pass on the value in parameters "upper" and "lower".
Refine the look by "w" and "arrow_lwd".}

\item{upper}{Size of the upper error bar.}

\item{lower}{Size of the lower error bar. By default, it equals the upper error bar.}

\item{arrow_width}{Width of the arrow head.}

\item{arrow_lwd}{Line width for the error bars.}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not
fit on the plot.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}
}
\description{
Create and save bar plots as .pdf, in "OutDir". If mdlink = TRUE, it inserts a .pdf and a .png
link in the markdown report, set by "path_of_report". The .png version is not created, only the
link is put in place, not to overwrite previous versions.
}
\examples{
MyVec <- 1:3
wbarplot(
  variable = MyVec, col = "gold1", sub = FALSE, w = 7, width = 1,
  incrBottMarginBy = 0, mdlink = FALSE, tilted_text = FALSE, hline = FALSE, vline = FALSE,
  filtercol = 1, lty = 1, lwd = 2, lcol = 2, errorbar = FALSE, upper = 0,
  arrow_width = 0.1, arrow_lwd = 1
)
}
