% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wbarplot}
\alias{wbarplot}
\title{wbarplot}
\usage{
wbarplot(variable, ..., col = "gold1", sub = F,
  plotname = substitute(variable), main = plotname, tilted_text = F,
  ylimits = NULL, hline = F, vline = F, filtercol = 1, lty = 1,
  lwd = 2, lcol = 2, errorbar = F, upper = 0, lower = upper,
  arrow_width = 0.1, arrow_lwd = 1, savefile = T, w = 7, h = w,
  incrBottMarginBy = 0, mdlink = b.mdlink)
}
\arguments{
\item{variable}{The variable to plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{col}{Color of the plot.}

\item{sub}{Subtitle below the plot.}

\item{plotname}{The name of the file saved.}

\item{main}{The title of the plot.}

\item{tilted_text}{Use 45 degree x-labels if TRUE. Useful for long, but not too many labels.}

\item{ylimits}{Defines the Y axis range. Replacement for the standard "ylim" argument.}

\item{hline}{Draw a horizontal line at the value you pass on to it. Useful to display a threshold. Design the line by "lty", "lwd" & "lcol" parameters.}

\item{vline}{Draw a vertical line at the value you pass on to it. Useful to display a threshold. Design the line by "lty", "lwd" & "lcol" parameters.}

\item{filtercol}{Color bars below / above the threshold with red / green. Define the direction by -1 or 1. Takes effect if "hline" is defined.}

\item{lty}{Linetype, defined by numbers 1-6.}

\item{lwd}{Linewidth. Set to 2 by default.}

\item{lcol}{Color of the line.}

\item{errorbar}{Draw error bars if TRUE. Pass on the value in parameters "upper" and "lower". Refine the look by "w" and "arrow_lwd".}

\item{upper}{Size of the upper error bar.}

\item{lower}{Size of the lower error bar. By default, it equals the upper error bar.}

\item{arrow_width}{Width of the arrow head.}

\item{arrow_lwd}{Line width for the error bars.}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not fit on the plot.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}
}
\description{
Create and save bar plots as .pdf, in "OutDir". If mdlink =T, it inserts a .pdf and a .png link in the markdown report, set by "path_of_report". The .png version is not created, only the link is put in place, not to overwrite previous versions.
}
\examples{
wbarplot (variable =  , ... =  , col = gold1, sub = F, plotname = substitute(variable), main = substitute(variable), w = 7, h = w, incrBottMarginBy = 0, mdlink = F, tilted_text = F, hline = F, vline = F, filtercol = 1, lty = 1, lwd = 2, lcol = 2, errorbar = F, upper = 0, lower = upper, arrow_width = 0.1, arrow_lwd = 1)
}
