% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wbarplot_dfCol}
\alias{wbarplot_dfCol}
\title{wbarplot_dfCol}
\usage{
wbarplot_dfCol(df, colName, col = "gold1", savefile = T, w = 7, h = 7,
  ...)
}
\arguments{
\item{df}{Input data frame to be plotted}

\item{col}{Color of the plot.Name}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{col}{Color of the plot.}
}
\description{
wbarplot for a column of a data frame.
}
\examples{
wbarplot_dfCol (df =  , colName =  , col = gold1, w = 7, h = 7, ... =  )
}

