% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wboxplot}
\alias{wboxplot}
\title{wboxplot}
\usage{
wboxplot(
  yourlist,
  main = as.character(substitute(yourlist)),
  sub = FALSE,
  ylab = "",
  col = get0("b.def.color", ifnotfound = "gold1"),
  incrBottMarginBy = 0,
  tilted_text = FALSE,
  savefile = get0("b.save.wplots", ifnotfound = TRUE),
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  mdlink = MarkdownHelpers::ww.set.mdlink(),
  PNG = get0("b.usepng", ifnotfound = TRUE),
  ...
)
}
\arguments{
\item{yourlist}{The variable to plot.}

\item{main}{Title of the plot and also the name of the file.}

\item{sub}{Subtitle below the plot.}

\item{ylab}{Y axis label}

\item{col}{Color of the plot.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not
fit on the plot.}

\item{tilted_text}{Use 45 degree x-labels if TRUE. Useful for long, but not too many labels.}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}
}
\description{
Create and save box plots as .pdf, in "OutDir". If mdlink = TRUE, it inserts a .pdf and a .png
link in the markdown report, set by "path_of_report". The .png version is not created, only the
link is put in place, not to overwrite previous versions.
}
\examples{
MyList <- list(rnorm(100), rnorm(100))
wboxplot(
  yourlist = MyList, col = "gold1",
  sub = FALSE, incrBottMarginBy = 0,
  tilted_text = FALSE, w = 7, mdlink = FALSE
)
}
