% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wcolorize}
\alias{wcolorize}
\title{wcolorize

Generate color palettes. Input: a vector with categories, can be numbers or strings.
Handles repeating values. Output: color vector of equal length as input.
Optionally it can ouput a list where an extra element lists the
categories (simply using unique would remove the names). See example.
Some color scale depend on packages "colorRamps", or "gplots".}
\usage{
wcolorize(
  vector = c(1, 1, 1:6),
  RColorBrewerSet = FALSE,
  ReturnCategoriesToo = FALSE,
  show = FALSE,
  randomize = FALSE,
  set = c(FALSE, "rich", "heat.colors", "terrain.colors", "topo.colors", "matlab",
    "rainbow")[1]
)
}
\arguments{
\item{vector}{A vector with categories, can be numbers or strings}

\item{RColorBrewerSet}{Use one of the RColorBrewer color sets? Provide that name}

\item{ReturnCategoriesToo}{Return unique Categories. See example.}

\item{show}{Show generated color palette}

\item{randomize}{Randomize colors}

\item{set}{Color palette for base
("heat.colors", "terrain.colors", "topo.colors", "rainbow"),
or "rich" for gplots::rich.colors, or  "matlab" for colorRamps::matlab.like.}
}
\description{
wcolorize

Generate color palettes. Input: a vector with categories, can be numbers or strings.
Handles repeating values. Output: color vector of equal length as input.
Optionally it can ouput a list where an extra element lists the
categories (simply using unique would remove the names). See example.
Some color scale depend on packages "colorRamps", or "gplots".
}
\examples{
wcolorize (vector = c(1, 1, 1:6), ReturnCategoriesToo = TRUE, show = TRUE)
}
