% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{whist.back2back}
\alias{whist.back2back}
\title{whist.back2back
Two back-to-back histograms from a list. The X-axis is only correct if  breaks1 ==breaks2. Undeveloped function, contains graphical bugs, no support for this function.}
\usage{
whist.back2back(ListOf2 = list(A = rnorm(100), B = rnorm(100)),
  breaks1 = 20, breaks2 = breaks1, colorz = c("green", "blue"), ...,
  plotname = substitute(variable), main_ = plotname, ylab = "Frequency",
  savefile = T, incrBottMarginBy = 0, w = 7, h = w, mdlink = F)
}
\arguments{
\item{ListOf2}{List of 2 numeric vectors}

\item{breaks1}{break parameter for histogram function for the 2st list element.}

\item{breaks2}{break parameter for histogram function for the 2st list element.}

\item{colorz}{Color of the 2 histograms}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{plotname}{The name of the file saved.}

\item{main_}{The title of the plot.}

\item{ylab}{Y-axis label}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not fit on the plot.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}
}
\description{
whist.back2back
Two back-to-back histograms from a list. The X-axis is only correct if  breaks1 ==breaks2. Undeveloped function, contains graphical bugs, no support for this function.
}
\examples{
whist.back2back(ListOf2 = list("A"  = rnorm(100), "B"=rnorm(100)))
}
