% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wlegend}
\alias{wlegend}
\title{wlegend}
\usage{
wlegend(
  NamedColorVec = NA,
  poz = 4,
  legend,
  cex = 0.75,
  bty = "n",
  ...,
  w = 7,
  h = w,
  title = NULL,
  ttl.by.varname = FALSE,
  OverwritePrevPDF = unless.specified("b.save.wplots"),
  mdlink = FALSE
)
}
\arguments{
\item{NamedColorVec}{Color of the boxes next to the text}

\item{poz}{Position of the legend (def: 4). Use numbers 1-4 to choose from
"topleft", "topright", "bottomright", "bottomleft".}

\item{legend}{Labels displayed (Text)}

\item{cex}{font size}

\item{bty}{The type of box to be drawn around the legend.
The allowed values are "o" (the default) and "n".}

\item{...}{Additional parameters for legend()}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{title}{What should be the title of the legend? NULL by default}

\item{ttl.by.varname}{Should the title of the legend substituted from the NamedColorVec variable's name?
ALSE by default. Does not work if you pass on a list item like this: list$element}

\item{OverwritePrevPDF}{Save the plot immediately with the same name
the last wplot* function made (It is stored in plotnameLastPlot variable).}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report
, set by "path_of_report".}
}
\description{
Quickly add a legend to an existing plot, and save the plot immediately.
}
\examples{
try.dev.off(); x = cbind(a = rnorm(1:10), b = rnorm(10)); wplot(x)
LegendCols = 2:5; names(LegendCols) = LETTERS[1:4]
wlegend(NamedColorVec = LegendCols, poz = 1, w = 7, bty = "n", OverwritePrevPDF = TRUE)
}
