% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wlegend}
\alias{wlegend}
\title{wlegend
Add a legend, and save the plot immediately}
\usage{
wlegend(x = "bottomleft", legend, fill = NULL, ..., bty = "n",
  OverwritePrevPDF = T)
}
\arguments{
\item{x}{location of legend}

\item{legend}{Labels displayed (Text)}

\item{fill}{Color of the boxes next to the text}

\item{...}{Pass any other parameter of the corresponding text function (most of them should work).}

\item{bty}{Background of legend, transparent by default}

\item{OverwritePrevPDF}{Save the plot immediately with the same name the last wplot* function made (It is stored in plotnameLastPlot variable).}
}
\description{
wlegend
Add a legend, and save the plot immediately
}
\examples{
wlegend(...)
}
