% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wlegend}
\alias{wlegend}
\title{wlegend}
\usage{
wlegend(x = c("topleft", "topright", "bottomright", "bottomleft")[4],
  fill_ = NULL, legend = names(fill_), ..., w_ = 7, h_ = w_,
  bty = "n", OverwritePrevPDF = T)
}
\arguments{
\item{fill_}{Color of the boxes next to the text}

\item{legend}{Labels displayed (Text)}

\item{...}{Additional parameters for legend()}

\item{w_}{Width of the saved pdf image, in inches.}

\item{h_}{Height of the saved pdf image, in inches.}

\item{bty}{The type of box to be drawn around the legend. The allowed values are "o" (the default) and "n".}

\item{OverwritePrevPDF}{Save the plot immediately with the same name the last wplot* function made (It is stored in plotnameLastPlot variable).}
}
\description{
wlegend
}
\examples{
function(x="bottomleft", fill_ = NULL, legend = names(fill_), ..., w_=7, h_=w_, bty = "n", OverwritePrevPDF =T)
}
