% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wLinRegression}
\alias{wLinRegression}
\title{wLinRegression}
\usage{
wLinRegression(DF, coeff = c("pearson", "spearman", "r2")[3],
  textlocation = "topleft", savefile = T, cexx = 1, ...)
}
\arguments{
\item{DF}{The same dataframe as you provided to wplot() before you called this function}

\item{coeff}{What coefficient to display? Either "pearson", "spearman" correlation values or "r2" for the Coefficient of Determination.}

\item{textlocation}{where to put the legend?}

\item{savefile}{Shall it call wplot_save_this(plotname = plotnameLastPlot) ?}

\item{cexx}{font size; 1 by default}

\item{...}{Additional parameters for the line to display.}
}
\description{
wLinRegression
}
\examples{
x = cbind(a=rnorm(1:10), b=rnorm(10)); wplot(x); wLinRegression(x, coeff = c("pearson", "spearman", "r2"))
}
