% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wpie}
\alias{wpie}
\title{wpie}
\usage{
wpie(
  NamedVector,
  percentage = TRUE,
  both_pc_and_value = FALSE,
  plotname = substitute(NamedVector),
  col = gplots::rich.colors(length(NamedVector)),
  savefile = get0("b.save.wplots", ifnotfound = TRUE),
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  mdlink = ww.set.mdlink(),
  PNG = get0("b.usepng", ifnotfound = FALSE),
  ...
)
}
\arguments{
\item{NamedVector}{The variable to plot.}

\item{percentage}{Display percentage instead of counts. TRUE by default.}

\item{both_pc_and_value}{Report both percentage AND number.}

\item{plotname}{Title of the plot (main parameter) and also the name of the file.}

\item{col}{Fill color. Defined by rich colors by default}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}
}
\description{
Create and save pie charts as .pdf, in "OutDir". If mdlink = TRUE, it inserts a .pdf and a .png
link in the markdown report, set by "path_of_report". The .png version is not created, only the
link is put in place, not to overwrite previous versions.
}
\examples{
Cake <- 1:3
names(Cake) <- letters[1:3]
wpie(Cake,
  percentage = TRUE,
  w = 7, mdlink = FALSE
)
}
