% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wpie}
\alias{wpie}
\title{wpie}
\usage{
wpie(variable, ..., percentage = TRUE, both_pc_and_value = F,
  plotname = substitute(variable),
  col = gplots::rich.colors(length(variable)), savefile = T, w = 7,
  h = w, mdlink = F)
}
\arguments{
\item{variable}{The variable to plot.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{percentage}{Display percentage instead of counts. TRUE by default.}

\item{both_pc_and_value}{Report both percentage AND number.}

\item{plotname}{Title of the plot (main parameter) and also the name of the file.}

\item{col}{Fill color. Defined by rich colours by default}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}
}
\description{
Create and save pie charts as .pdf, in "OutDir". If mdlink =T, it inserts a .pdf and a .png link in the markdown report, set by "path_of_report". The .png version is not created, only the link is put in place, not to overwrite previous versions.
}
\examples{
wpie (variable =  , ... =  , percentage = TRUE, plotname = substitute(variable), w = 7, h = w, mdlink = F)
}
