% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wplot}
\alias{wplot}
\title{wplot}
\usage{
wplot(
  df2col,
  col = 1,
  pch = 18,
  ...,
  panel_first = grid(NULL),
  plotname = substitute(df2col),
  errorbar = FALSE,
  upper = 0,
  lower = upper,
  left = 0,
  right = left,
  width.whisker = 0.1,
  arrow_lwd = 1,
  col_errorbar = 1,
  ylim = FALSE,
  xlim = FALSE,
  abline = c(FALSE, "v", "h", "ab")[1],
  a = FALSE,
  b = FALSE,
  lty = 1,
  lwd = 1,
  col_abline = 1,
  equal.axes = FALSE,
  savefile = get0("b.save.wplots", ifnotfound = TRUE),
  mdlink = MarkdownHelpers::ww.set.mdlink(),
  w = get0("b.defSize", ifnotfound = 7),
  h = w,
  PNG = get0("b.usepng", ifnotfound = FALSE)
)
}
\arguments{
\item{df2col}{Input data frame to be plotted (2 columns)}

\item{col}{Color of the plot.}

\item{pch}{Define the symbol for each data point. A number (0-25) or any string between ""-s.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}

\item{panel_first}{Draw a backround grid, if set to "grid(NULL)"}

\item{plotname}{Title of the plot (main parameter) and also the name of the file.}

\item{errorbar}{Draw error bars if TRUE. Pass on the value in parameters "upper" and "lower".
Refine the look by "w" and "arrow_lwd".}

\item{upper}{Size of the upper error bar.}

\item{lower}{Size of the lower error bar. By default, it equals the upper error bar.}

\item{left}{Size of the left error bar.}

\item{right}{Size of the right error bar. By default, it equals the left error bar.}

\item{width.whisker}{Width of the error bar whisker.}

\item{arrow_lwd}{Line width for the error bar arrow. Line width for the error bar arrow.}

\item{col_errorbar}{Color of the error bar arrow.}

\item{ylim}{manual Y-limits error bar}

\item{xlim}{manual X-limits error bar}

\item{abline}{Draw a line on the plot. Any value from: c( 'v', 'h', 'ab') for vertical,
horizontal, and line with any slope. In each case you need to specify "a = ", specifying the
X-position for vertical (v); the Y-position for horizontal (yh) lines and the intercept for
lines with a slope (ab). In the latter case, 'b' specifes the slope.}

\item{a}{X-offset for vertical lines, Y-offset for horizontal, and inclined lines.}

\item{b}{Slope of an inclined line.}

\item{lty}{Linetype, defined by numbers 1-6.}

\item{lwd}{Line width. Set to 2 by default.}

\item{col_abline}{Color of the line.}

\item{equal.axes}{Span of axes is set to equal (maximum range in either X or Y).}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}
}
\description{
Create and save scatter plots as .pdf, in "OutDir". If mdlink = TRUE, it inserts a .pdf and a .png
link in the markdown report, set by "path_of_report". The .png version is not created, only the
link is put in place, not to overwrite previous versions. The .png version is not created, only
the link is put in place. You can add 2D error bars around the dots, or add lines (ablines) to
your plot, by setting "abline" argument to = FALSE (no line, default), "h" (horizontal, further
specified by a = y-offset), "v" (vertical, further specified by a = x-offset), "ab" (line with an
angle, further specified by a = offset, b = slope).
}
\examples{
try.dev.off()
mydf <- cbind("A" = rnorm(100), "B" = rpois(100, 8))
wplot(
  df2col = mydf, col = 1, pch = 18, w = 7,
  mdlink = FALSE, errorbar = FALSE, upper = 0,
  left = 0, right = left, width.whisker = 0.1, arrow_lwd = 1, abline = FALSE,
  a = FALSE, b = FALSE, lty = 1, lwd = 1, col_abline = 1
)
}
