% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wplot_save_pheatmap}
\alias{wplot_save_pheatmap}
\title{wplot_save_pheatmap}
\usage{
wplot_save_pheatmap(
  x,
  data = NULL,
  suffix = "heatmap",
  plotname = as.character(substitute(x)),
  add = 1,
  width = max(if (is.null(data)) 7 else ncol(data) + add, 4),
  height = max(if (is.null(data)) 7 else nrow(data) - add, 4),
  pdf = TRUE,
  png = FALSE,
  png_res = 100,
  png_dim_factor = 100,
  mdlink = TRUE
)
}
\arguments{
\item{x}{The pheatmap object to save.}

\item{data}{data matrix used to create the pheatmap object. Only used to define plot height and width.}

\item{suffix}{Suffix to File name. Default: 'heatmap'.}

\item{plotname}{Name of the plot. Default: substitute(x). File saved as .pdf, inside working
directory.}

\item{add}{Constant to add to width and height. Default: 2.}

\item{width}{width of the plot in inches.}

\item{height}{height of the plot in inches.}

\item{pdf}{Save as pdf. Default: TRUE.}

\item{png}{Save as png. Default: F.}

\item{png_res}{Super unintuitive way to control plot "outcome" (works super unpredictably).}

\item{png_dim_factor}{Width is in inches for pdf and pixels for png. This is a multiplier factor. Default: 100.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}
}
\description{
Save pheatmap object. Modified from:
https://stackoverflow.com/questions/43051525/how-to-draw-pheatmap-plot-to-screen-and-also-save-to-file
}
\examples{
test <- matrix(rnorm(200), 20, 10)
colnames(test) <- paste("Test", 1:10, sep = "")
rownames(test) <- paste("Gene", 1:20, sep = "")
ph.test <- pheatmap::pheatmap(test)
wplot_save_pheatmap(ph.test)
}
