% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wplot_save_pheatmap}
\alias{wplot_save_pheatmap}
\title{wplot_save_pheatmap}
\usage{
wplot_save_pheatmap(
  x,
  suffix = "heatmap",
  plotname = substitute(x),
  width = 15,
  height = width,
  pdf = TRUE,
  png = FALSE,
  png_dim_factor = 100,
  mdlink = TRUE
)
}
\arguments{
\item{x}{The pheatmap object to save.}

\item{suffix}{Suffix to File name. Default: 'heatmap'.}

\item{width}{width of the plot in inches.}

\item{height}{height of the plot in inches.}

\item{pdf}{Save as pdf. Default: TRUE.}

\item{png}{Save as png. Default: F.}

\item{png_dim_factor}{Width is in inches for pdf and pixels for png. This is a multiplier factor. Default: 100.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{filename}{File name (saved as .pdf, inside working directory).}
}
\description{
Save pheatmap object. Modified from:
https://stackoverflow.com/questions/43051525/how-to-draw-pheatmap-plot-to-screen-and-also-save-to-file
}
\examples{
test = matrix(rnorm(200), 20, 10);
colnames(test) = paste("Test", 1:10, sep = "");
rownames(test) = paste("Gene", 1:20, sep = "");
ph.test <- pheatmap::pheatmap(test);
wplot_save_pheatmap(ph.test)
}
