% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wplot_save_this}
\alias{wplot_save_this}
\title{wplot_save_this}
\usage{
wplot_save_this(plotname = date(), ..., w = 7, h = 7, mdlink = FALSE)
}
\arguments{
\item{plotname}{Title of the plot (main parameter) and also the name of the file.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}
}
\description{
Save the currently active graphic device (for complicated plots).  Insert links to your markdown report, set by "path_of_report". Name the file by naming the variable!
}
\examples{
wplot_save_this (plotname = date(), col = gold1, ... =  , w = 7, h = 7, mdlink = FALSE, ManualName = FALSE)
}

