% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wscatter.fill}
\alias{wscatter.fill}
\title{wscatter.fill}
\usage{
wscatter.fill(x, y, ..., color, xlim = range(x), ylim = range(y),
  zlim = range(color), nlevels = 20, pch = 21, cex = 1,
  plotname = substitute(variable), plot.title = plotname,
  xlb = substitute(x), ylb = substitute(y), plot.axes, key.title, key.axes,
  asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, savefile = T, w = 7, h = w, incrBottMarginBy = 0,
  mdlink = F)
}
\arguments{
\item{x}{X variable}

\item{y}{Y variable}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{color}{Filling color of the symbols}

\item{xlim}{Manually set the range of canvas in X dimension}

\item{zlim}{Manually set the range of colors numbers (Z dimension)}

\item{nlevels}{Number of steps in the color gradient}

\item{pch}{Define the symbol for each data point. A number [0-25] or any string between ""-s.}

\item{cex}{Size of the symbols}

\item{plotname}{The name of the file saved.}

\item{plot.title}{The title of the plot.}

\item{plot.axes}{Draw axis ticks}

\item{key.title}{...}

\item{key.axes}{...}

\item{asp}{numeric, giving the aspect ratio y/x. See help('plot.window').}

\item{xaxs}{The style of axis interval calculation to be used for the X-axis. See help('par').}

\item{yaxs}{The style of axis interval calculation to be used for the X-axis. See help('par').}

\item{las}{numeric in {0,1,2,3}; the style of axis labels. See help('par').}

\item{axes}{Draw axes and box}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not fit on the plot.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{ylimManually}{set the range of canvas in Y dimension}
}
\description{
wscatter.fill
}
\examples{
wbarplot (variable =  , ... =  , col = gold1, sub = F, plotname = substitute(variable), main = substitute(variable), w = 7, h = w, incrBottMarginBy = 0, mdlink = F, tilted_text = F, hline = F, vline = F, filtercol = 1, lty = 1, lwd = 2, lcol = 2, errorbar = F, upper = 0, lower = upper, arrow_width = 0.1, arrow_lwd = 1)
wscatter.fill(x=rnorm(100), y=rnorm(100), color=rnorm(100), nlevels=15, pch = 21, xlab="The X Dimension. Wooaaahh")
}
