% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wstripchart}
\alias{wstripchart}
\title{wstripchart}
\usage{
wstripchart(
  yourlist,
  main = as.character(substitute(yourlist)),
  sub = NULL,
  ylab = "",
  BoxPlotWithMean = FALSE,
  border = 1,
  incrBottMarginBy = 0,
  tilted_text = FALSE,
  metod = "jitter",
  jitter = 0.3,
  pch = 18,
  pchlwd = 1,
  cex.lab = 1,
  pchcex = 1.5,
  bg = "seagreen2",
  colorbyColumn = TRUE,
  col = if (colorbyColumn) {
     1:length(yourlist)
 } else {
     1
 },
  savefile = unless.specified("b.save.wplots"),
  w = unless.specified("b.defSize", 7),
  h = w,
  mdlink = ww.set.mdlink(),
  PNG = unless.specified("b.usepng", F),
  ...
)
}
\arguments{
\item{yourlist}{Input list to be plotted.}

\item{main}{Title of the plot (main parameter) and also the name of the file.}

\item{sub}{Subtitle below the plot.}

\item{ylab}{Y axis label}

\item{BoxPlotWithMean}{Display the mean instead of the median in a boxplot. This is non-standard
use of a boxplot, report it.}

\item{border}{An optional vector of colors for the outlines of the boxplots. The values in border
are recycled if the length of border is less than the number of plots.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not
fit on the plot.}

\item{tilted_text}{Use 45 degree x-labels if TRUE. Useful for long, but not too many labels.}

\item{metod}{Method for displaying data points to avoid overlap; either"jitter" or "stack". See
stripchart().}

\item{jitter}{The amount of horizontal scatter added to the individual data points (to avoid
overlaps).}

\item{pch}{Define the symbol for each data point. A number (0-25) or any string between ""-s.}

\item{pchlwd}{Define the outline width of the symbol for each data point.}

\item{cex.lab}{Cex for labels}

\item{pchcex}{Define the size of the symbol for each data point.}

\item{bg}{Background color.}

\item{colorbyColumn}{Color each box by a simple background color? TRUE by default.}

\item{col}{Color of the plot.}

\item{savefile}{Save plot as pdf in OutDir, TRUE by default.}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by
"path_of_report".}

\item{PNG}{Set to true if you want to save the plot as PNG instead of the default PDF.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should
work).}
}
\description{
Create and save strip charts as .pdf, in "OutDir". If mdlink = TRUE, it inserts a .pdf and a .png
link in the markdown report, set by "path_of_report". The .png version is not created, only the
link is put in place, not to overwrite previous versions.
}
\examples{
try.dev.off()
my.ls <- list(A = rnorm(10), B = rnorm(10), C = rnorm(10))
wstripchart(yourlist = my.ls)
}
