% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{wstripchart_list}
\alias{wstripchart_list}
\title{wstripchart_list}
\usage{
wstripchart_list(yalist, ..., plotname = as.character(substitute(yalist)),
  sub = FALSE, ylb = "NULL", xlab = "NULL", border = 1, bxpcol = 0,
  pch = 23, pchlwd = 1, pchcex = 1.5, bg = "chartreuse2",
  coll = "black", metod = "jitter", jitter = 0.2, w = 7, h = 7,
  incrBottMarginBy = 0, tilted_text = F, mdlink = F)
}
\arguments{
\item{yalist}{Input list to be plotted.}

\item{...}{Pass any other parameter of the corresponding plotting function (most of them should work).}

\item{plotname}{Title of the plot (main parameter) and also the name of the file.}

\item{sub}{Subtitle for the plot.}

\item{ylb}{Y-axis label.}

\item{border}{An optional vector of colors for the outlines of the boxplots. The values in border are recycled if the length of border is less than the number of plots.}

\item{bxpcol}{Color of the boxplot outlines.}

\item{pch}{Define the symbol for each data point. A number [0-25] or any string between ""-s.}

\item{pchlwd}{Define the outline width of the symbol for each data point.}

\item{pchcex}{Define the size of the symbol for each data point.}

\item{bg}{Background color.}

\item{metod}{Method for displaying data points to avoid overlap; either"jitter" or "stack". See stripchart().}

\item{jitter}{The amount of horizontal scatter added to the individual data points (to avoid overlaps).}

\item{w}{Width of the saved pdf image, in inches.}

\item{h}{Height of the saved pdf image, in inches.}

\item{incrBottMarginBy}{Increase the blank space at the bottom of the plot. Use if labels do not fit on the plot.}

\item{tilted_text}{Use 45 degree x-labels if TRUE. Useful for long, but not too many labels.}

\item{mdlink}{Insert a .pdf and a .png image link in the markdown report, set by "path_of_report".}

\item{xlb}{X-axis label.}

\item{col}{Color of the plot.l}
}
\description{
Create and save stripcharts from a list as .pdf, in "OutDir". This version allows individual coloring of each data point, by a color-list of the same dimension. If mdlink =T, it inserts a .pdf and a .png link in the markdown report, set by "path_of_report". The .png version is not created, only the link is put in place, not to overwrite previous versions.
}
\examples{
wstripchart_list (yalist =  , ... =  , plotname = as.character(substitute(yalist)), sub = FALSE, ylb = NULL, xlab = NULL, border = 1, bxpcol = 0, pch = 23, pchlwd = 1, pchcex = 1.5, bg = chartreuse2, coll = black, metod = jitter, jitter = 0.2, w = 7, h = 7, incrBottMarginBy = 0, tilted_text = F, mdlink = F)
}

