% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.dev.copy}
\alias{ww.dev.copy}
\title{ww.dev.copy}
\usage{
ww.dev.copy(PNG_ = FALSE, PNG_res = 450, w_, h_, fname_)
}
\arguments{
\item{PNG_}{Set to true if you want to save the plot as PNG instead of the default PDF.}

\item{PNG_res}{deafult 450}

\item{w_}{Width of the saved pdf image, in inches.}

\item{h_}{Height of the saved pdf image, in inches.}

\item{fname_}{File name}
}
\description{
Parser for dev.copy to save as PDF or PNG
}
\examples{
try.dev.off(); plot(1); # ww.dev.copy(PNG = FALSE, w_ = 7, h_ = 7, fname_ = "myNewplot")
}
