% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww.set.OutDir}
\alias{ww.set.OutDir}
\title{ww.set.OutDir}
\usage{
ww.set.OutDir(dir = OutDir)
}
\arguments{
\item{dir}{OutDir to check and set.}
}
\description{
Checks if global variable OutDir is defined. If not,
it returns the current working directory
}
\examples{
ww.set.OutDir()
}
