% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkdownReports.R
\name{ww_autoPlotName}
\alias{ww_autoPlotName}
\title{ww_autoPlotName}
\usage{
ww_autoPlotName(name = NULL)
}
\arguments{
\item{name}{Manually name your plot}
}
\description{
Internal function. Creates automatic plot and file-names.
}
\examples{
ww_autoPlotName()
}
